% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassScores.R
\name{binaryClassScores}
\alias{binaryClassScores}
\title{Compare the true classification with a matrix of predictions}
\usage{
binaryClassScores(y, predictMatrix)
}
\arguments{
\item{y}{True labels of the observations. Ordering of y must match that of the predictMatrix.}

\item{predictMatrix}{Predicted labels of the observations.}
}
\description{
Compare the true classification with a matrix of predictions
}
\examples{
x = iris[51:150, -5]
y = as.factor(iris[51:150, 5])
tmp2 = svmCV_multi(x = x, y = y,
                   nFolds = 5, nExp = 100, cores = 1)

predictMatrix = purrr::map(tmp2, "svmPredictOrderedVector") \%>\%
  purrr::map(as.character) \%>\%
  do.call(rbind,.)
binaryClassScores(y = y, predictMatrix = predictMatrix)
}
\author{
Kevin Wang
}
