% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_cor.R
\name{calcIdenDist}
\alias{calcIdenDist}
\title{Identity Distance for a matrix}
\usage{
calcIdenDist(matrix)
}
\arguments{
\item{matrix}{a matrix}
}
\description{
Identity Distance for a matrix
}
\examples{
set.seed(1)
x = matrix(rnorm(100), ncol = 10)
calcIdenDist(x)
}
