% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_liners.R
\name{clean_digits}
\alias{clean_digits}
\title{Clean the digits in a data frame}
\usage{
clean_digits(df, digits = 2, method = c("round", "signif"))
}
\arguments{
\item{df}{A data frame with numbers and categorical variables}

\item{digits}{Number of significant figures}

\item{method}{Either "round" or "signif"}
}
\value{
A data frame with less digits
}
\description{
Clean the digits in a data frame
}
\author{
Kevin Wang
}
