% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names_char.R
\name{clean_names_char}
\alias{clean_names_char}
\title{Clean a character vector in the same algorithm as janitor::clean_names}
\usage{
clean_names_char(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A cleaned character vector
}
\description{
Clean a character vector in the same algorithm as janitor::clean_names
}
