% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDigits.R
\name{cleanDigits}
\alias{cleanDigits}
\title{Clean the digits in a data frame}
\usage{
cleanDigits(df, sig = 2)
}
\arguments{
\item{df}{A data frame with numbers and categorical variables}

\item{sig}{Number of significant figures}
}
\value{
A data frame with less digits
}
\description{
Clean the digits in a data frame
}
