% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareBinaryClassResults.R
\name{compareBinaryClassResults}
\alias{compareBinaryClassResults}
\title{Compare the true classification with a matrix of predictions}
\usage{
compareBinaryClassResults(y, classifierMatrix)
}
\arguments{
\item{y}{True labels of the observations. Ordering of y must match that of the predictMatrix.}

\item{classifierMatrix}{Classifiers are rows for each column being observations}
}
\description{
Compare the true classification with a matrix of predictions
}
\examples{
x = iris[51:150, -5] + rnorm(100*4)
y = as.factor(iris[51:150, 5])
tmp1 = svmCV_multi(x = x, y = y,
                   nFolds = 10, nExp = 100, cores = 1)

tmp2 = svmCV_multi(x = x, y = y,
                   nFolds = 5, nExp = 100, cores = 1)

predictMatrix1 = purrr::map(tmp1, "svmPredictOrderedVector") \%>\%
  purrr::map(as.character) \%>\%
  do.call(rbind,.)

predictMatrix2 = purrr::map(tmp2, "svmPredictOrderedVector") \%>\%
  purrr::map(as.character) \%>\%
  do.call(rbind,.)


classifierMatrix = rbind(
  binaryClassScores(y = y, predictMatrix = predictMatrix1),
  binaryClassScores(y = y, predictMatrix = predictMatrix2)
)
rownames(classifierMatrix) = c("TenFold", "FiveFold")

compareBinaryClassResults(y, classifierMatrix)#'
}
