% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confTable.R
\name{confTable}
\alias{confTable}
\title{Confusion table with forced levels}
\usage{
confTable(x, y)
}
\arguments{
\item{x}{a logical vector}

\item{y}{a logical vector}
}
\description{
Confusion table with forced levels
}
\examples{
set.seed(11)
x = sample(c(TRUE), 100, replace = TRUE)
y = sample(c(TRUE, FALSE), 100, replace = TRUE)
table(x, y)
confTable(x, y)
}
\author{
Kevin Wang
}
