% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confTable.R
\name{confTableStats}
\alias{confTableStats}
\title{Confusion table statistics from caret}
\usage{
confTableStats(x, y)
}
\arguments{
\item{x}{a logical vector}

\item{y}{a logical vector}
}
\description{
Confusion table statistics from caret
}
\examples{
set.seed(11)
x = sample(c(TRUE), 100, replace = TRUE)
y = sample(c(TRUE, FALSE), 100, replace = TRUE)
confTableStats(x, y)
}
\author{
Kevin Wang
}
