% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_multi.R
\name{cv_multi}
\alias{cv_multi}
\title{Performing CV using random forest, allowing multiple cores}
\usage{
cv_multi(x, y, method = "svm", nfolds = 5, nexp = 5)
}
\arguments{
\item{x}{design matrix}

\item{y}{factor}

\item{method}{Method of the classification algorithm}

\item{nfolds}{number of CV folds}

\item{nexp}{number of loops}
}
\description{
Performing CV using random forest, allowing multiple cores
}
\examples{
x = iris[51:150, -5]
y = factor(iris[51:150, 5])
res = cv_multi(x = x, y = y, method = "rf")
\dontrun{
library(furrr)
plan(multisession(workers = 5))
res = cv_multi(x = x, y = y, method = "rf", nexp = 20)
plan(future::sequential)
}
}
\author{
Kevin Wang
}
