% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvPartition.R
\name{cvPartition}
\alias{cvPartition}
\title{Creating CV data partition}
\usage{
cvPartition(x, y, nFolds)
}
\arguments{
\item{x}{x should be a matrix without rownames}

\item{y}{y should be true labels}

\item{nFolds}{number of folds}
}
\description{
Creating CV data partition
}
\examples{
x = iris[51:150, -5]
y = factor(iris[51:150, 5])
cvp = cvPartition(x = x, y = y, nFolds = 5)
str(cvp)
}
\author{
Kevin Wang
}
