% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_roxygenise.R
\name{execute_roxygenise}
\alias{execute_roxygenise}
\title{Execute roxygen2::roxygenise().}
\usage{
execute_roxygenise()
}
\description{
Call this function to send "roxygen2::roxygenise()" to
the console of rstudio and execute.
Steps:
1. Install the KW package.
2. Go to Tools -> Addins -> Browse Addins in a RStudio session.
3. Create a shortcut for this function.
}
