% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLassoCoef.R
\name{getLassoCoef}
\alias{getLassoCoef}
\title{Compare the true classification with a matrix of predictions}
\usage{
getLassoCoef(lassoObj, s)
}
\arguments{
\item{lassoObj}{Lasso object, either cv.glmnet or glmnet outputs}

\item{s}{}
}
\description{
Compare the true classification with a matrix of predictions
}
\examples{
x = iris[51:150, -5] \%>\% as.matrix
y = as.matrix(as.integer(iris[51:150, 5])-2)
library(glment)
lassoObj = glmnet(x = x, y = y, family = "binomial")
getLassoCoef(lassoObj, s = 0.001)
cvLassoObj = cv.glmnet(x = x, y = y, family = "binomial")
getLassoCoef(cvLassoObj, s = "lambda.min")
}
\author{
Kevin Wang
}
