% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLassoPath.R
\name{ggLassoPath}
\alias{ggLassoPath}
\title{ggplot for lasso path plot}
\usage{
ggLassoPath(lassoObj, colour)
}
\arguments{
\item{lassoObj}{is a glmnet object from the `glmnet` package}

\item{colour}{is a vector that is of the same length as the number of variables fed into `lassoObj`}
}
