% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_explorer.R
\name{ggplot_explorer}
\alias{ggplot_explorer}
\title{Shiny app for ggplot exploration}
\usage{
ggplot_explorer(uploaded_plot)
}
\arguments{
\item{g}{a ggplot object}
}
\description{
Shiny app for ggplot exploration
}
\examples{
if(interactive()){
uploaded_plot = ggplot(data = iris, aes(x = Sepal.Length, y = Sepal.Width,
           colour = Species)) +
  geom_point()
ggplot_explorer(uploaded_plot = uploaded_plot)
}
}
\author{
Kevin Wang
}
