% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_cor.R
\name{identityDist}
\alias{identityDist}
\title{Identity Distance}
\usage{
identityDist(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\description{
Identity Distance
}
\examples{
set.seed(1)
x = rnorm(100)
identityDist(x = x, y = x)
identityDist(x = x, y = 2*x)
}
