% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_roxygenise.R
\name{insert_roxygenise}
\alias{insert_roxygenise}
\title{Insert roxygen2::roxygenise().}
\usage{
insert_roxygenise()
}
\description{
Call this function as an addin to insert \code{ roxygen2::roxygenise() } at the cursor position.
}
