% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoCV.R
\name{lassoCV}
\alias{lassoCV}
\title{Performing CV using lasso}
\usage{
lassoCV(cvObj, family, s)
}
\arguments{
\item{cvObj}{outputs of the cvPartition function}
}
\description{
Performing CV using lasso
}
\examples{
x = iris[51:150, -5] \%>\% as.matrix
y = as.matrix(as.integer(iris[51:150, 5])-2)
library(glmnet)
cvp = cvPartition(x = x, y = y, nFolds = 5)
lassoCV(cvp, family = "binomial", s = "lambda.min")

}
\author{
Kevin Wang
}
