% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoCV_multi.R
\name{lassoCV_multi}
\alias{lassoCV_multi}
\title{Performing CV using lasso, allowing multiple cores}
\usage{
lassoCV_multi(x, y, family, s = "lambda.min", nFolds = 5, nExp,
  cores = 1)
}
\arguments{
\item{cvObj}{outputs of the cvPartition function}
}
\description{
Performing CV using lasso, allowing multiple cores
}
\examples{
x = iris[51:150, -5] \%>\% as.matrix
y = as.matrix(as.integer(iris[51:150, 5])-2)
library(glmnet)
lassoCvObj = lassoCV_multi(x = x, y = y, family = "binomial", s = "lambda.min", nExp = 10)
purrr::map_dbl(lassoCvObj, "cvLassoMeanError") \%>\% mean
}
\author{
Kevin Wang
}
