% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_cv.R
\name{loess_cv}
\alias{loess_cv}
\title{Calculate loess span parameter using cross validation}
\usage{
loess_cv(formula, data, span = seq(0.1, 1, by = 0.1), v = 5, repeats = 2)
}
\arguments{
\item{formula}{see loess()}

\item{data}{see loess()}

\item{span}{see loess()}

\item{v}{Number of CV fold, see rsample::vfold_cv()}

\item{repeats}{NUmber of CV repeats, see rsample::vfold_cv()}
}
\description{
Calculate loess span parameter using cross validation
}
\examples{
library(dplyr)
library(ggplot2)

data(economics, package="ggplot2")  # load data
economics$index <- 1:nrow(economics)  # create index variable
data <- economics[1:80, ]  # retail 80rows for better graphical understanding
lresult = loess_cv(uempmed ~ index, data = data,
span = c(0.1, 0.2, 0.3, 0.4, 0.5), v = 5, repeats = 20)

lresult \%>\%
group_by(span) \%>\%
summarise(sse = mean(resid2))

lresult \%>\%
ggplot(aes(x = factor(span), y = resid2)) +
geom_boxplot()
}
