% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitCV.R
\name{logitCV}
\alias{logitCV}
\title{Performing CV using glm(family = "binomial")}
\usage{
logitCV(cvObj, cutoff = 0.5)
}
\arguments{
\item{cvObj}{outputs of the cvPartition function}
}
\description{
Performing CV using glm(family = "binomial")
}
\examples{
x = iris[51:150, -5] + rnorm(100, 0, 2)
y = ifelse(iris[51:150, 5] == "versicolor", 0, 1)
cvObj = cvPartition(x = x, y = y, nFolds = 5)
table(pred = logitCV(cvObj)$logitPredictIntOrderedVector, y)

}
