% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitCV_multi.R
\name{logitCV_multi}
\alias{logitCV_multi}
\title{Performing CV using glm(family = "binomial"), allowing multiple cores}
\usage{
logitCV_multi(x, y, nFolds, nExp, cores = 1, cutoff = 0.5)
}
\arguments{
\item{cvObj}{outputs of the cvPartition function}
}
\description{
Performing CV using glm(family = "binomial"), allowing multiple cores
}
\examples{
x = iris[51:150, -5] + rnorm(100, 0, 2)
y = ifelse(iris[51:150, 5] == "versicolor", 0, 1)
tmp2 = logitCV_multi(x = x, y = y, nFolds = 5, nExp = 10, cores = 1)

purrr::map_dbl(tmp2, "logitMeanError")
}
