% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedVector2Df.R
\name{namedVector2Df}
\alias{namedVector2Df}
\title{Convert a named vector into a data.frame}
\usage{
namedVector2Df(vec, names = names("variable", "value"))
}
\arguments{
\item{vec}{A named vector}
}
\description{
Convert a named vector into a data.frame
}
