% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_liners.R
\name{padded_names}
\alias{padded_names}
\title{Sequence of variable names, all with the same length}
\usage{
padded_names(prefix, p)
}
\arguments{
\item{prefix}{Prefix of the variable names}

\item{p}{Number of variables}
}
\description{
Sequence of variable names, all with the same length
}
\examples{
paste0("X", 1:10)
padded_names("X", 10)
}
