% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_panel.R
\name{panel.cor}
\alias{panel.cor}
\title{A better pairs plot function}
\usage{
panel.cor(x, y)
}
\arguments{
\item{data}{a data frame}

\item{subset}{A subsetting index}
}
\description{
A better pairs plot function
}
\examples{
pairs_cor(iris[,-5])
}
