% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_cor.R
\name{panel_cor}
\alias{panel_cor}
\title{A better pairs plot function}
\usage{
panel_cor(x, y)
}
\arguments{
\item{data}{a data frame}
}
\description{
A better pairs plot function
}
\examples{
pairs(iris[,-5], lower.panel = panel_cor)
}
\author{
Kevin Wang
}
