% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_scatter_abline.R
\name{panel_scatter_abline}
\alias{panel_scatter_abline}
\title{A better pairs plot function}
\usage{
panel_scatter_abline(x, y, a = 0, b = 1, col = "red")
}
\arguments{
\item{data}{a data frame}
}
\description{
A better pairs plot function
}
\examples{
pairs(iris[,-5],
lower.panel = panel_cor,
upper.panel = panel_scatter_abline)
}
