% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{Plotting PCA via a prcomp object}
\usage{
pcaPlot(pcaObj, obsData, colVar, ellipse = F, topFeatures = 5,
  arrows = F, label.size = 2)
}
\arguments{
\item{pcaObj}{prcomp object}

\item{obsData}{a data.frame matching the number of rows of the PCA data}

\item{colVar}{Character vector corresponding to one of the columns of obsData}

\item{ellipse}{Should ellipse be plotted based on colVar}

\item{topFeatures}{Number of original variables most associated with each of the PCs}

\item{arrows}{Should arrows be plotted}

\item{label.size}{Size of the label for the arrows}
}
\value{
A ggplot
}
\description{
Plotting PCA via a prcomp object
}
\author{
Kevin Wang
}
