% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLE.R
\name{plot_RLE}
\alias{plot_RLE}
\title{RLE plot}
\usage{
plot_RLE(exprsMatrix)
}
\arguments{
\item{exprsMatrix}{a matrix}
}
\description{
RLE plot
}
\examples{
n = 30
p = 1000
exprsMatrix = cbind(
matrix(rnorm((n - 3)*p), nrow = p),
matrix(rnorm(3*p, mean = 5), nrow = p)
)
boxplot(exprsMatrix)
plot_RLE(exprsMatrix)
}
\author{
Kevin Wang
}
