% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlm.R
\name{plotlm}
\alias{plotlm}
\title{Plot scatter plot of two vectors and fit a linear model}
\usage{
plotlm(x, y, se = FALSE)
}
\arguments{
\item{x}{x vector}

\item{y}{y vector}
}
\description{
Plot scatter plot of two vectors and fit a linear model
}
\examples{
set.seed(10)
x = rnorm(100)
y = rnorm(100)
plotlm(x = x, y = y)
}
