% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_partition.R
\name{print.cv_obj}
\alias{print.cv_obj}
\title{Printing cv_partition object}
\usage{
\method{print}{cv_obj}(cv_obj)
}
\arguments{
\item{cv_obj}{Output from cv_partition}
}
\description{
Printing cv_partition object
}
