% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_partition.R
\name{print.cv_pred_result}
\alias{print.cv_pred_result}
\title{Printing cv_partition object}
\usage{
\method{print}{cv_pred_result}(cv_pred_result)
}
\arguments{
\item{cv_pred_result}{Output from *_cv functions}
}
\description{
Printing cv_partition object
}
