% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfCV.R
\name{rfCV}
\alias{rfCV}
\title{Performing CV using random forest}
\usage{
rfCV(cvObj, ntree)
}
\arguments{
\item{cvObj}{outputs of the cvPartition function}
}
\description{
Performing CV using random forest
}
\author{
Kevin Wang
}
