% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfCV_multi.R
\name{rfCV_multi}
\alias{rfCV_multi}
\title{Performing CV using random forest, allowing multiple cores}
\usage{
rfCV_multi(x, y, nFolds = 5, nExp, cores = 1, ntree = 500)
}
\arguments{
\item{x}{design matrix}

\item{y}{factor}

\item{nFolds}{number of CV folds}

\item{nExp}{number of loops}

\item{cores}{number of cores to perform the CV}

\item{ntree}{number of trees}
}
\description{
Performing CV using random forest, allowing multiple cores
}
\examples{
x = iris[51:150, -5]
y = factor(iris[51:150, 5])
tmp2 = rfCV_multi(x = x, y = y,
                   nFolds = 5, nExp = 100, cores = 1, ntree = 100)

purrr::map_dbl(tmp2, "rfMeanError")
purrr::map(tmp2, "rfVoteOrderedMatrix")
}
\author{
Kevin Wang
}
