% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplewiseCor.R
\name{samplewiseCor}
\alias{samplewiseCor}
\title{Calculate samplwise correlation}
\usage{
samplewiseCor(x, y, method = "pearson")
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix, potentially has more columns than x}

\item{method}{See methods in cor}
}
\description{
Calculate samplwise correlation
}
