% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_scotch.R
\name{scale_fill_scotch}
\alias{scale_fill_scotch}
\title{Scotch colours.}
\usage{
scale_fill_scotch(stata = FALSE, ...)
}
\arguments{
\item{n}{The number of colours}
}
\value{
ggproto object
}
\description{
Scotch colours.
}
\examples{
df = data.frame(
x = rnorm(100),
y = rnorm(100),
colour = factor(1:10))
p = ggplot(df, aes(x = x, y = y, colour = colour)) +
geom_point()
p + scale_colour_scotch()

}
