% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotchColours.R
\name{scotchColour}
\alias{scotchColour}
\title{Better colour palette than Set 1 from library(RColorBrewer)}
\usage{
scotchColour(stata)
}
\arguments{
\item{n}{The number of colours needed from Set 1 from library(RColorBrewer)}
}
\value{
Returns a function of colours
}
\description{
Better colour palette than Set 1 from library(RColorBrewer)
}
\examples{
scales::show_col(scotchColour()(10))
}
