% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_sub_front_last.R
\name{str_sub_front_last}
\alias{str_sub_front_last}
\title{Substring based on number of first and last characters}
\usage{
str_sub_front_last(string, front, last, sep = "...")
}
\arguments{
\item{string}{string}

\item{front}{number of characters at the front}

\item{last}{number of characters at the back}

\item{sep}{separator in the middle}
}
\value{
Returns a function of colours
}
\description{
Substring based on number of first and last characters
}
\examples{
string = c("Apples", "Banana")
front = 2
last = 2
str_sub_front_last(string, front, last)
}
\author{
Kevin Wang
}
