% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_sub_frontLast.R
\name{str_sub_frontLast}
\alias{str_sub_frontLast}
\title{Substring based on number of first and last characters}
\usage{
str_sub_frontLast(string, frontNum, lastNum, sep = "...")
}
\arguments{
\item{string}{string}

\item{frontNum}{number of characters at the front}

\item{lastNum}{number of characters at the back}

\item{sep}{separator in the middle}
}
\value{
Returns a function of colours
}
\description{
Substring based on number of first and last characters
}
\examples{
string = c("Apples", "Banana")
frontNum = 2
lastNum = 2
str_sub_frontLast(string, frontNum, lastNum)
string = c("Apples", "Apnaes")
str_sub_frontLast(string, frontNum, lastNum)
}
\author{
Kevin Wang
}
