% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_cv.R
\name{svm_cv}
\alias{svm_cv}
\title{Performing CV using svm}
\usage{
svm_cv(cvObj)
}
\arguments{
\item{cv_obj}{outputs of the cvPartition function}
}
\description{
Performing CV using svm
}
\examples{
x = iris[51:150, -5]
y = factor(iris[51:150, 5])
cv_obj = cv_partition(x = x, y = y, nfolds = 5)
svm_cv(cv_obj)
}
\author{
Kevin Wang
}
