% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmCV.R
\name{svmCV}
\alias{svmCV}
\title{Performing CV using svm}
\usage{
svmCV(cvObj)
}
\arguments{
\item{cvObj}{outputs of the cvPartition function}
}
\description{
Performing CV using svm
}
\examples{
x = iris[50:150, -5]
y = iris[50:150, 5] \%>\% as.factor
cvp = cvPartition(x = x, y = y, nFolds = 5)
table(pred = svmCV(cvp)$svmPredictOrderedVector, y)

}
