% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmCV_multi.R
\name{svmCV_multi}
\alias{svmCV_multi}
\title{Performing CV using svm, allowing multiple cores}
\usage{
svmCV_multi(x, y, nFolds = 5, nExp, cores = 1)
}
\arguments{
\item{cvObj}{outputs of the cvPartition function}
}
\description{
Performing CV using svm, allowing multiple cores
}
\examples{
x = iris[51:150, -5]
y = iris[51:150, 5] \%>\% as.factor
tmp2 = svmCV_multi(x = x, y = y,
                   nFolds = 5, nExp = 100, cores = 1)

purrr::map_dbl(tmp2, "svmMeanError")
}
\author{
Kevin Wang
}
