% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2DataFrame.R
\name{table2DataFrame}
\alias{table2DataFrame}
\title{Turn a table into a data frame}
\usage{
table2DataFrame(x, variable.name = "variable.name",
  value.name = "value.name")
}
\arguments{
\item{x}{a frequency table (i.e. only one row)}

\item{variable.name}{a character}

\item{value.name}{a character}
}
\description{
Turn a table into a data frame
}
\author{
Kevin Wang
}
