% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{v2df}
\alias{v2df}
\alias{t2df}
\title{Convert a named vector into a data.frame}
\usage{
v2df(vec, variable_name = "variable", value_name = "value")

t2df(table, variable_name = "variable", value_name = "value")
}
\arguments{
\item{vec}{A named vector}

\item{variable_name}{variable_name}

\item{value_name}{value_name}

\item{table}{a frequency table (i.e. only one row)}
}
\description{
Convert a named vector into a data.frame

Turn a one-row table into a data frame
}
\examples{
x = 1:5
names(x) = letters[1:5]
v2df(x)
x = table(sample(letters[1:5], 100, replace = TRUE))
t2df(x)
}
\author{
Kevin Wang
}
