% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_liners.R
\name{top}
\alias{top}
\title{Display the first k rows and first k columns of the data}
\usage{
top(x, k = 6)
}
\arguments{
\item{x}{data.frame}

\item{k}{number of rows and columns to view}
}
\description{
Display the first k rows and first k columns of the data
}
\examples{
top(mtcars)
}
\author{
Kevin Wang
}
