% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topView.R
\name{topView}
\alias{topView}
\title{Display the first k rows and first k columns of the data}
\usage{
topView(x, k = 6)
}
\arguments{
\item{x}{data.frame}

\item{k}{number of rows and columns to view}
}
\description{
Display the first k rows and first k columns of the data
}
\author{
Kevin Wang
}
