% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsnePlot.R
\name{tsnePlot}
\alias{tsnePlot}
\title{Plotting tSNE via a Rtsne object}
\usage{
tsnePlot(RtsneObj, obsData, colVar, ellipse = T)
}
\arguments{
\item{RtsneObj}{Rtsne object}

\item{obsData}{a data.frame matching the number of rows of the PCA data}

\item{colVar}{Character vector corresponding to one of the columns of obsData}

\item{ellipse}{Should ellipse be plotted based on colVar}
}
\description{
Plotting tSNE via a Rtsne object
}
\author{
Kevin Wang
}
