% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDigits.R
\name{cleanDigits}
\alias{cleanDigits}
\title{Add together two numbers}
\usage{
cleanDigits(df, sig = 2)
}
\arguments{
\item{df}{A data frame with numbers and categorical variables}

\item{sig}{Number of significant figures}
}
\value{
A data frame with less digits
}
\description{
Add together two numbers
}
\examples{
cleanDigits(mtcars, 1)




}
