% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvPartition.R
\name{cvPartition}
\alias{cvPartition}
\title{Creating CV data partition}
\usage{
cvPartition(x, y, nFolds)
}
\arguments{
\item{x}{x should be a matrix without rownames}

\item{y}{y should be true labels
x = iris[50:150, -5]
y = iris[50:150, 5] %>% as.factor
cvp = cvPartition(x = x, y = y, nFolds = 5)
str(cvp)}
}
\description{
Creating CV data partition
}
