% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_panel.R
\name{panel.cor}
\alias{panel.cor}
\title{A better pairs plot function}
\usage{
panel.cor(x, y, subset)
}
\arguments{
\item{subset}{A subsetting index}

\item{data}{a data frame}
}
\description{
A better pairs plot function
}
\examples{
subset = sample(c(TRUE,FALSE), 150, replace = TRUE)
pairs_cor(iris[,-5], subset = subset)
}
