% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_cv_multi.R
\name{rf_cv_multi}
\alias{rf_cv_multi}
\alias{rf_cv}
\title{Performing CV using random forest, allowing multiple cores}
\usage{
rf_cv_multi(x, y, nfolds = 5, nexp = 5)

rf_cv(cv_obj)
}
\arguments{
\item{x}{design matrix}

\item{y}{factor}

\item{nfolds}{number of CV folds}

\item{nexp}{number of loops}

\item{cv_obj}{outputs of the cvPartition function}
}
\description{
Performing CV using random forest, allowing multiple cores

Performing CV using random forest
}
\examples{
x = iris[51:150, -5]
y = factor(iris[51:150, 5])
res = rf_cv_multi(x = x, y = y)
\dontrun{
library(furrr)
plan(multisession(workers = 5))
res = rf_cv_multi(x = x, y = y, nexp = 20)
plan(future::sequential)
}
x = iris[51:150, -5]
y = factor(iris[51:150, 5])
cv_obj = cv_partition(x = x, y = y, nfolds = 5)
rf_cv(cv_obj)
}
\author{
Kevin Wang

Kevin Wang
}
