% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotchColours.R
\name{scotchColour}
\alias{scotchColour}
\title{Better colour palette than Set 1 from library(RColorBrewer)}
\usage{
scotchColour(n)
}
\arguments{
\item{n}{The number of colours needed from Set 1 from library(RColorBrewer)}
}
\description{
Better colour palette than Set 1 from library(RColorBrewer)
}
\examples{
df = data.frame(
x = rnorm(100),
y = rnorm(100),
colour = factor(1:10))
ggplot(df, aes(x = x, y = y, colour = colour)) +
geom_point() +
scale_colour_manual(values = scotchColour(n = df$colour \%>\% unique \%>\% length))

}
