% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmCV.R
\name{svmCV}
\alias{svmCV}
\title{Performing CV using svm}
\usage{
svmCV(cvObj)
}
\arguments{
\item{cvObj}{outputs of the cvPartition function}
}
\description{
Performing CV using svm
}
\examples{
x = iris[1:100, -5]
y = 2*(iris[1:100, 5] \%>\% as.factor \%>\% as.integer()) - 3L
s = sample(1:100)
x = x[s,]
y = y[s]
tmp = cvPartition(x = x, y = y, nFolds = 5)
str(tmp)
table(pred = svmCV(tmp)$svmPredictOrderedVector, y)
}
