% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmCV_multi.R
\name{svmCV_multi}
\alias{svmCV_multi}
\title{Performing CV using svm, allowing multiple cores}
\usage{
svmCV_multi(x, y, nFolds, nExp, cores = 1)
}
\arguments{
\item{cvObj}{outputs of the cvPartition function}
}
\description{
Performing CV using svm, allowing multiple cores
}
\examples{

x = iris[50:150, -5]
}
