% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{arcsinh}
\alias{arcsinh}
\alias{expit}
\title{Arcsinh transform}
\usage{
arcsinh(x)

expit(x)
}
\arguments{
\item{x}{A vector, but also takes matrix and purely numeric data.frame}
}
\description{
Arcsinh transform

The expit function
}
\examples{
set.seed(11)
x = rnorm(100)
plot(x, arcsinh(x))
curve(expit, from = -5, to = 5)
}
\author{
Kevin Wang
}
